+++ internal/service/categories.go
@@ -8,11 +8,11 @@ import (
 
 type CategoriesService interface {
 	Create(categoryDTO *domain.CategoryDTO) (*domain.CategoryDTO, error)
-	GetAll(userID int) (*[]domain.CategoryDTO, error)
-	GetByName(name string, userID int) (*domain.CategoryDTO, error)
-	GetByEmoji(emoji string, userID int) (*domain.CategoryDTO, error)
-	GetByID(ID int, userID int) (*domain.CategoryDTO, error)
-	DeleteByID(id int, userID int) error
+	GetAll(userID int64) (*[]domain.CategoryDTO, error)
+	GetByName(name string, userID int64) (*domain.CategoryDTO, error)
+	GetByEmoji(emoji string, userID int64) (*domain.CategoryDTO, error)
+	GetByID(ID int64, userID int64) (*domain.CategoryDTO, error)
+	DeleteByID(id int64, userID int64) error
 	UpdateByID(categoryDTO *domain.CategoryDTO) (*domain.CategoryDTO, error)
 }
 type categoriesService struct {
@@ -36,7 +36,7 @@ func (s *categoriesService) Create(categoryDTO *domain.CategoryDTO) (*domain.Cat
 
 	return category.ToDTO(), nil
 }
-func (s *categoriesService) GetAll(userID int) (*[]domain.CategoryDTO, error) {
+func (s *categoriesService) GetAll(userID int64) (*[]domain.CategoryDTO, error) {
 	categories, err := s.repo.GetAll(userID)
 	if err != nil {
 		return nil, err
@@ -49,7 +49,7 @@ func (s *categoriesService) GetAll(userID int) (*[]domain.CategoryDTO, error) {
 
 	return &categoryDTOs, nil
 }
-func (s *categoriesService) GetByName(name string, userID int) (*domain.CategoryDTO, error) {
+func (s *categoriesService) GetByName(name string, userID int64) (*domain.CategoryDTO, error) {
 	name = sanitizer.Sanitize(name)
 
 	category, err := s.repo.GetByName(name, userID)
@@ -59,7 +59,7 @@ func (s *categoriesService) GetByName(name string, userID int) (*domain.Category
 
 	return category.ToDTO(), nil
 }
-func (s *categoriesService) GetByEmoji(emoji string, userID int) (*domain.CategoryDTO, error) {
+func (s *categoriesService) GetByEmoji(emoji string, userID int64) (*domain.CategoryDTO, error) {
 	category, err := s.repo.GetByEmoji(emoji, userID)
 	if err != nil {
 		return nil, err
@@ -67,7 +67,7 @@ func (s *categoriesService) GetByEmoji(emoji string, userID int) (*domain.Catego
 
 	return category.ToDTO(), nil
 }
-func (s *categoriesService) DeleteByID(id int, userID int) error {
+func (s *categoriesService) DeleteByID(id int64, userID int64) error {
 	return s.repo.DeleteByID(id, userID)
 }
 func (s *categoriesService) UpdateByID(categoryDTO *domain.CategoryDTO) (*domain.CategoryDTO, error) {
@@ -82,7 +82,7 @@ func (s *categoriesService) UpdateByID(categoryDTO *domain.CategoryDTO) (*domain
 
 	return category.ToDTO(), nil
 }
-func (s *categoriesService) GetByID(ID int, userID int) (*domain.CategoryDTO, error) {
+func (s *categoriesService) GetByID(ID int64, userID int64) (*domain.CategoryDTO, error) {
 	category, err := s.repo.GetByID(ID, userID)
 	if err != nil {
 		return nil, err
